+++ src/pages/Max50ListPage/Max50ListPage.css
@@ -13,22 +13,3 @@
   max-width: 200px;
   max-height: 100px;
 }
-.searchs {
-  width: 100%;
-  height: 12vh;
-  box-shadow: rgb(150, 174, 255) 0px 1px 10px;
-  background-color: white;
-  overflow: hidden;
-  overflow-y: auto;
-}
-.listItem {
-  width: auto;
-  height: auto;
-  display: flex;
-  align-items: center;
-  padding-top: 2%;
-  padding-left: 2%;
-}
-.listItem:hover {
-  background-color: lightgray;
-}
