+++ src/pages/SearchBar/SearchBar.css
@@ -1,0 +1,19 @@
+.searchs {
+  width: 100%;
+  height: 12vh;
+  box-shadow: rgb(150, 174, 255) 0px 1px 10px;
+  background-color: white;
+  overflow: hidden;
+  overflow-y: auto;
+}
+.listItem {
+  width: auto;
+  height: auto;
+  display: flex;
+  align-items: center;
+  padding-top: 2%;
+  padding-left: 2%;
+}
+.listItem:hover {
+  background-color: lightgray;
+}
